+++ CHANGES.md
@@ -32,6 +32,7 @@
   * Wii: Unencrypted images from RVT-H Reader systems are now supported.
   * NES: Improved publisher lookup for FDS titles.
   * Wii U: Display game publishers.
+  * DMG: Added more entry point formats.
 
 * New compressed texture formats:
   * Ericsson ETC1 and ETC2